리뷰 계산 로직 수정

+++ src/main/java/com/petweb/sponge/trainer/dto/ReviewCreate.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 public class ReviewCreate {
 
     private Long trainerId;
-    private float score;
+    private int score;
     private String content;
 }
